# src/swagger.json
{
  "openapi" : "3.0.0",
  "servers" : [ {
    "description" : "sayori",
    "url" : "https://acow-api-dev.sayori.link"
  } ],
  "info" : {
    "description" : "This is a simple API",
    "version" : "1.0.0",
    "title" : "ACOW API",
    "contact" : {
      "email" : "you@your-company.com"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
...
  "paths" : {
...
    "/seasons/{seasonId}/contract-types" : {
      "parameters" : [ {
        "schema" : {
          "type" : "string"
        },
        "name" : "seasonId",
        "in" : "path",
        "required" : true,
        "description" : "season id"
      } ],
      "get" : {
        "summary" : "Retrieves a season's available contract types",
        "tags" : [ ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "success" : {
                      "type" : "boolean",
                      "x-stoplight" : {
                        "id" : "2oyno6x3d4ijb"
                      }
                    },
                    "contractTypes" : {
                      "type" : "array",
                      "x-stoplight" : {
                        "id" : "s8akyjivw0wzo"
                      },
                      "items" : {
                        "$ref" : "#/components/schemas/ContractType"
                      }
                    }
                  }
                },
                "examples" : {
                  "Example 1" : {
                    "value" : {
                      "success" : true,
                      "contractTypes" : [ {
                        "id" : "1",
                        "name" : "Base Contract",
                        "slug" : "base",
                        "icon" : "kirby",
                        "discordChannelId" : "1077741836749242468",
                        "assignmentStart" : "2026-05-15T00:00:00.000Z",
                        "assignmentEnd" : "2026-05-16T00:00:00.000Z",
                        "reviewDeadline" : "2026-05-23T00:00:00.000Z"
                      } ]
                    }
                  }
                }
              }
            }
          }
        },
        "operationId" : "get-seasons-seasonId-contract-types"
      },
      "post" : {
        "summary" : "[Admin] Create a new contract type for a season",
        "tags" : [ ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "success" : {
                      "type" : "boolean",
                      "x-stoplight" : {
                        "id" : "zd4he2whpi8t1"
                      }
                    },
                    "contractType" : {
                      "$ref" : "#/components/schemas/ContractType"
                    }
                  }
                },
                "examples" : {
                  "Example 1" : {
                    "value" : {
                      "success" : true,
                      "contractType" : {
                        "id" : "1",
                        "name" : "Dating Contract",
                        "slug" : "dating",
                        "icon" : "hearts",
                        "discordChannelId" : "1077741836749242468",
                        "assignmentStart" : "2026-05-15T00:00:00.000Z",
                        "assignmentEnd" : "2026-05-16T00:00:00.000Z",
                        "reviewDeadline" : "2026-05-23T00:00:00.000Z",
                        "createdAt" : "2026-05-14T00:00:00.000Z",
                        "updatedAt" : "2026-05-14T00:00:00.000Z"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "operationId" : "post-seasons-seasonId-contract-types",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ContractTypeEditable"
              },
              "examples" : {
                "Example 1" : {
                  "value" : {
                    "name" : "Base Contract",
                    "slug" : "base",
                    "icon" : "kirby",
                    "discordChannelId" : "1077741836749242468",
                    "assignmentStart" : "2026-05-15T00:00:00.000Z",
                    "assignmentEnd" : "2026-05-16T00:00:00.000Z",
                    "reviewDeadline" : "2026-05-23T00:00:00.000Z"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
...
      "ContractType" : {
        "type" : "object",
        "x-examples" : {
          "Example 1" : {
            "id" : "1",
            "name" : "Base Contract",
            "slug" : "base",
            "icon" : "kirby",
            "discordChannelId" : "1077741836749242468",
            "assignmentStart" : "2026-05-15T00:00:00.000Z",
            "assignmentEnd" : "2026-05-16T00:00:00.000Z",
            "reviewDeadline" : "2026-05-23T00:00:00.000Z"
          }
        },
        "properties" : {
          "id" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "slug" : {
            "type" : "string"
          },
          "icon" : {
            "type" : "string"
          },
          "discordChannelId" : {
            "type" : "string"
          },
          "assignmentStart" : {
            "type" : "string"
          },
          "assignmentEnd" : {
            "type" : "string"
          },
          "reviewDeadline" : {
            "type" : "string",
            "x-stoplight" : {
              "id" : "5wz0tdmqtnlqm"
            }
          },
          "createdAt" : {
            "type" : "string",
            "x-stoplight" : {
              "id" : "xl8m59bagv5xb"
            }
          },
          "updatedAt" : {
            "type" : "string",
            "x-stoplight" : {
              "id" : "wz0hc4x7ivdub"
            }
          }
        }
      },
      "ContractTypeEditable" : {
        "type" : "object",
        "x-examples" : {
          "Example 1" : {
            "id" : "1",
            "name" : "Base Contract",
            "slug" : "base",
            "icon" : "kirby",
            "discordChannelId" : "1077741836749242468",
            "assignmentStart" : "2026-05-15T00:00:00.000Z",
            "assignmentEnd" : "2026-05-16T00:00:00.000Z",
            "reviewDeadline" : "2026-05-23T00:00:00.000Z"
          }
        },
        "properties" : {
          "name" : {
            "type" : "string"
          },
          "slug" : {
            "type" : "string"
          },
          "icon" : {
            "type" : "string"
          },
          "discordChannelId" : {
            "type" : "string"
          },
          "reviewDeadline" : {
            "type" : "string",
            "x-stoplight" : {
              "id" : "dvtfcmtmidoti"
            }
          },
          "assignmentStart" : {
            "type" : "string"
          },
          "assignmentEnd" : {
            "type" : "string"
          }
        }
      }
    },
    "securitySchemes" : {
      "access_token" : {
        "name" : "Authorization",
        "type" : "apiKey",
        "in" : "header"
      }
    }
  },
  "security" : [ {
    "access_token" : [ ]
  } ]
}